+++ pywbemtools/pywbemcli/_context_obj.py
@@ -314,7 +314,7 @@ class ContextObj(object):
         try:
             if self.pdb:
                 import pdb  # pylint: disable=import-outside-toplevel
-                pdb.set_trace()  # pylint: disable=no-member
+                pdb.set_trace()  # pylint: disable=forgotten-debug-statement
 
             cmd()  # The pywbemcli command function call.